Move worker pool to shared/concurrency/
