+++ flow/pkg/go.mod
@@ -13,8 +13,6 @@ require (
 	github.com/stretchr/testify v1.12.1
 	go.mongodb.org/mongo-driver/v2 v2.8.0
 	go.temporal.io/sdk v1.47.0
-	go.uber.org/goleak v1.3.0
-	golang.org/x/sync v0.22.0
 	google.golang.org/api v0.287.1
 )
 
@@ -95,6 +93,7 @@ require (
 	golang.org/x/mod v0.37.0 // indirect
 	golang.org/x/net v0.57.0 // indirect
 	golang.org/x/oauth2 v0.36.0 // indirect
+	golang.org/x/sync v0.22.0 // indirect
 	golang.org/x/sys v0.47.0 // indirect
 	golang.org/x/telemetry v0.0.0-20260625142307-59b4966ccb57 // indirect
 	golang.org/x/text v0.40.0 // indirect
