+++ flow/shared/concurrency/workers_test.go
@@ -1,4 +1,4 @@
-package common
+package concurrency
 
 import (
 	"context"
